requirements.txt
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
tqdm==4.66.4
